Bugfix: don't allow user to submit empty string for newPersonName
This fixes an issue where a user could submit an empty string for newPersonName in the ModalPersonEdit component from the FaceDashboard.

I did this myself and now I can't access the `/faces` route because FaceComponent expects cell.image to always be a string.

Any pointers on how to fix that issue with the data on my docker hosted instance would be appreciated.